refactor(buildx): add buildx driver during configure-docker
