+++ pkg/cmd/root/root.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/depot/cli/pkg/buildx/commands"
 	bakeCmd "github.com/depot/cli/pkg/cmd/bake"
 	buildCmd "github.com/depot/cli/pkg/cmd/build"
 	cacheCmd "github.com/depot/cli/pkg/cmd/cache"
@@ -54,8 +53,7 @@ func NewCmdRoot(version, buildDate string) *cobra.Command {
 	cmd.AddCommand(loginCmd.NewCmdLogin())
 	cmd.AddCommand(logout.NewCmdLogout())
 	cmd.AddCommand(versionCmd.NewCmdVersion(version, buildDate))
-	cmd.AddCommand(dockerCmd.NewCmdConfigureDocker())
-	cmd.AddCommand(commands.NewBuildxCmd(dockerCli))
+	cmd.AddCommand(dockerCmd.NewCmdConfigureDocker(dockerCli))
 
 	return cmd
 }
